// src/pages/AuthPage/ErrorDisplay/types.ts
export enum ErrorOptionsDisplay {
    ErrorRetry = 'errorRetry',
    ErrorLimitSMS = 'errorLimitSMS',
    ErrorDefault = 'errorDefault',
    ErrorMessageDontSend = 'errorMessageDontSend',
    ErrorPhoneNotFound = 'errorPhoneNotFound',
    ErrorRejected = 'errorRejected',
    ErrorSimCardIncorect = 'errorSimCardIncorect',
    ErrorSMSDisabled = 'errorSMSDisabled',
    ErrorWarningMessage = 'errorWarningMessage',
}